refactor: セレクトボックスをコンポーネントに切り出し、設定値のディレクトリ名を変更
