--- src/configs/selectorInfo.ts
@@ -1,15 +1,0 @@
-export const selectorInfo = [
-  { value: undefined, label: 'calories' },
-  { value: undefined, label: 'protein' },
-  { value: undefined, label: 'fat' },
-  { value: undefined, label: 'sodium' },
-  { value: undefined, label: 'fiber' },
-  { value: undefined, label: 'carbo' },
-  { value: undefined, label: 'sugars' },
-  { value: undefined, label: 'potass' },
-  { value: undefined, label: 'vitamins' },
-  { value: undefined, label: 'shelf' },
-  { value: undefined, label: 'weight' },
-  { value: undefined, label: 'cups' },
-  { value: undefined, label: 'rating' },
-];
