+++ src/constants/chartInfo.ts
